refactor(api): move certification point of contact route
in identity-access-management
