+++ api/tests/identity-access-management/integration/application/user/user.route.test.js
@@ -1,3 +1,4 @@
+import { certificationPointOfContactController } from '../../../../../lib/application/certification-point-of-contacts/certification-point-of-contact-controller.js';
 import { identityAccessManagementRoutes } from '../../../../../src/identity-access-management/application/routes.js';
 import { config } from '../../../../../src/shared/config.js';
 import * as i18nPlugin from '../../../../../src/shared/infrastructure/plugins/i18n.js';
@@ -130,6 +131,29 @@ describe('Integration | Identity Access Management | Application | Route | User'
         // then
         expect(response.statusCode).to.equal(403);
       });
+    });
+  });
+
+  describe('GET /api/certification-point-of-contacts/me', function () {
+    it('returns controller success response HTTP code', async function () {
+      // given
+      const userId = databaseBuilder.factory.buildUser().id;
+      const headers = {
+        authorization: generateValidRequestAuthorizationHeader(userId),
+      };
+      sinon.stub(certificationPointOfContactController, 'get').callsFake((request, h) => h.response().code(200));
+
+      // when
+      const result = await httpTestServer.request(
+        'GET',
+        '/api/certification-point-of-contacts/me',
+        null,
+        null,
+        headers,
+      );
+
+      // then
+      expect(result.statusCode).to.equal(200);
     });
   });
 });
